+++ CHANGELOG.md
@@ -17,6 +17,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * Added deprecation API `core/meta/deprecation`
 * Added `Object.isPromiseLike`
+* Added `Then.prototype.finally`
 
 ## v3.0.0-beta.94 (2019-12-18)